+++ mapping_library/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.5] 25 Oktober 2019
+* Added fixed objects layer
+* Added a basic scalebar as Fixed object
+* Some bugfixes and example changes
+
 ## [0.0.4] 22 Oktober 2019
 * Some bugfixes
 * Code formatting
@@ -11,13 +16,11 @@
 * Updated example application
 
 ## [0.0.2] 3 Oktober 2019
-
 * Format changes
 * Added sample
 * Started on API Documentation
 
 ## [0.0.1] 2 Oktober 2019
-
 * TODO: Describe initial release.